Move qubitization_walk_operator.py to a new module qubitization/
