+++ docs/bloqs/index.rst
@@ -84,7 +84,7 @@ Bloqs Library
     qft/approximate_qft.ipynb
     phase_estimation/lp_resource_state.ipynb
     phase_estimation/text_book_qpe.ipynb
-    qubitization_walk_operator.ipynb
+    qubitization/qubitization_walk_operator.ipynb
     phase_estimation/qubitization_qpe.ipynb
 
 .. toctree::
