+++ qualtran/bloqs/chemistry/thc/walk_operator.py
@@ -16,7 +16,7 @@ import numpy as np
 from numpy.typing import NDArray
 
 from qualtran.bloqs.chemistry.thc import PrepareTHC, SelectTHC
-from qualtran.bloqs.qubitization_walk_operator import QubitizationWalkOperator
+from qualtran.bloqs.qubitization.qubitization_walk_operator import QubitizationWalkOperator
 
 
 def get_walk_operator_for_thc_ham(
