+++ qualtran/serialization/resolver_dict.py
@@ -98,7 +98,7 @@ import qualtran.bloqs.qft.qft_phase_gradient
 import qualtran.bloqs.qft.qft_text_book
 import qualtran.bloqs.qft.two_bit_ffft
 import qualtran.bloqs.qsp.generalized_qsp
-import qualtran.bloqs.qubitization_walk_operator
+import qualtran.bloqs.qubitization.qubitization_walk_operator
 import qualtran.bloqs.reflection
 import qualtran.bloqs.reflection_using_prepare
 import qualtran.bloqs.rotations.hamming_weight_phasing
@@ -301,7 +301,7 @@ RESOLVER_DICT = {
     "qualtran.bloqs.qft.qft_text_book.QFTTextBook": qualtran.bloqs.qft.qft_text_book.QFTTextBook,
     "qualtran.bloqs.qft.two_bit_ffft.TwoBitFFFT": qualtran.bloqs.qft.two_bit_ffft.TwoBitFFFT,
     "qualtran.bloqs.qsp.generalized_qsp.GeneralizedQSP": qualtran.bloqs.qsp.generalized_qsp.GeneralizedQSP,
-    "qualtran.bloqs.qubitization_walk_operator.QubitizationWalkOperator": qualtran.bloqs.qubitization_walk_operator.QubitizationWalkOperator,
+    "qualtran.bloqs.qubitization.qubitization_walk_operator.QubitizationWalkOperator": qualtran.bloqs.qubitization.qubitization_walk_operator.QubitizationWalkOperator,
     "qualtran.bloqs.reflection.Reflection": qualtran.bloqs.reflection.Reflection,
     "qualtran.bloqs.reflection_using_prepare.ReflectionUsingPrepare": qualtran.bloqs.reflection_using_prepare.ReflectionUsingPrepare,
     "qualtran.bloqs.rotations.hamming_weight_phasing.HammingWeightPhasing": qualtran.bloqs.rotations.hamming_weight_phasing.HammingWeightPhasing,
